txs: share message as file from modal, module refactoring
